+++ Runtime/Scripts/AssetPreviewAttribute.cs
@@ -10,8 +10,8 @@ namespace EditorAttributes
 		/// <summary>
 		/// Attribute to preview an asset in the inspector
 		/// </summary>
-		/// <param name="previewWidth">The width of the preview</param>
-		/// <param name="previewHeight">The height of the preview</param>
+		/// <param name="previewWidth">The width of the preview in pixels</param>
+		/// <param name="previewHeight">The height of the preview in pixels/></param>
 		public AssetPreviewAttribute(float previewWidth = 0f, float previewHeight = 0f)
         {
             PreviewWidth = previewWidth;