+++ Runtime/Scripts/MessageBoxAttribute.cs
@@ -10,7 +10,7 @@ namespace EditorAttributes
         Error
     }
 
-    public class MessageBoxAttribute : PropertyAttribute, IConditionalAttribute
+    public class MessageBoxAttribute : PropertyAttribute, IConditionalAttribute, IDynamicStringAttribute
     {
 		public int EnumValue{ get; private set; }
 		public bool DrawProperty { get; private set; }
@@ -19,6 +19,7 @@ namespace EditorAttributes
 		public string ConditionName { get; private set; }
 
 		public MessageMode MessageType { get; private set; }
+		public StringInputMode StringInputMode { get; private set; }
 
 		/// <summary>
 		/// Attribute to display a message box depending on a condition
@@ -26,7 +27,8 @@ namespace EditorAttributes
 		/// <param name="message">The message to display</param>
 		/// <param name="conditionName">The condition to evaluate</param>
 		/// <param name="messageType">The type of the message</param>
-		public MessageBoxAttribute(string message, string conditionName, MessageMode messageType = MessageMode.Log)
+		/// <param name="stringInputMode">Set if the string input is set trough a constant or dynamically trough another member</param>
+		public MessageBoxAttribute(string message, string conditionName, MessageMode messageType = MessageMode.Log, StringInputMode stringInputMode = StringInputMode.Constant)
 #if UNITY_2023_3_OR_NEWER
         : base(true) 
 #endif
@@ -35,6 +37,7 @@ namespace EditorAttributes
 			ConditionName = conditionName;
 			DrawProperty = true;
 			MessageType = messageType;
+			StringInputMode = stringInputMode;
 		}
 
 		/// <summary>
@@ -44,7 +47,8 @@ namespace EditorAttributes
 		/// <param name="conditionName">The condition to evaluate</param>
 		/// <param name="enumValue">The value of the enum</param>
 		/// <param name="messageType">The type of the message</param>
-		public MessageBoxAttribute(string message, string conditionName, object enumValue, MessageMode messageType = MessageMode.Log)
+		/// <param name="stringInputMode">Set if the string input is set trough a constant or dynamically trough another member</param>
+		public MessageBoxAttribute(string message, string conditionName, object enumValue, MessageMode messageType = MessageMode.Log, StringInputMode stringInputMode = StringInputMode.Constant)
 #if UNITY_2023_3_OR_NEWER
         : base(true) 
 #endif
@@ -63,7 +67,8 @@ namespace EditorAttributes
 		/// <param name="conditionName">The condition to evaluate</param>
 		/// <param name="drawProperty">Draw the property this attribute is attached to</param>
 		/// <param name="messageType">The type of the message</param>
-		public MessageBoxAttribute(string message, string conditionName, bool drawProperty, MessageMode messageType = MessageMode.Log)
+		/// <param name="stringInputMode">Set if the string input is set trough a constant or dynamically trough another member</param>
+		public MessageBoxAttribute(string message, string conditionName, bool drawProperty, MessageMode messageType = MessageMode.Log, StringInputMode stringInputMode = StringInputMode.Constant)
 #if UNITY_2023_3_OR_NEWER
         : base(true) 
 #endif
@@ -82,7 +87,8 @@ namespace EditorAttributes
 		/// <param name="enumValue">The value of the enum</param>
 		/// <param name="drawProperty">Draw the property this attribute is attached to</param>
 		/// <param name="messageType">The type of the message</param>
-		public MessageBoxAttribute(string message, string conditionName, object enumValue, bool drawProperty, MessageMode messageType = MessageMode.Log)
+		/// <param name="stringInputMode">Set if the string input is set trough a constant or dynamically trough another member</param>
+		public MessageBoxAttribute(string message, string conditionName, object enumValue, bool drawProperty, MessageMode messageType = MessageMode.Log, StringInputMode stringInputMode = StringInputMode.Constant)
 #if UNITY_2023_3_OR_NEWER
         : base(true) 
 #endif