+++ Runtime/Scripts/PrefixAttribute.cs
@@ -13,7 +13,7 @@ namespace EditorAttributes
 		/// Attribute to add a prefix on a field
 		/// </summary>
 		/// <param name="prefix">The prefix to add</param>
-		/// <param name="offset">Offset to add between the prefix and field</param>
+		/// <param name="offset">Offset to add between the prefix and field in pixels</param>
 		/// <param name="stringInputMode">Set if the string input is set trough a constant or dynamically trough another member</param>
 		public PrefixAttribute(string prefix, float offset = 0f, StringInputMode stringInputMode = StringInputMode.Constant)
 		{