+++ Runtime/Scripts/ProgressBarAttribute.cs
@@ -11,7 +11,7 @@ namespace EditorAttributes
 		/// Attribute to draw a progress bar
 		/// </summary>
 		/// <param name="maxValue">The maximum value of the progress bar</param>
-		/// <param name="barHeight">The height of the progress bar</param>
+		/// <param name="barHeight">The height of the progress bar in pixels</param>
 		public ProgressBarAttribute(float maxValue = 100f, float barHeight = 20f)
 		{
 			MaxValue = maxValue;